requirements.txt
# API & Data Collection
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyspark==3.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Analysis & Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1